+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 - Save the state (open, height, active panel) on `localStorage`
 - Added `start!` function for easier configuration
 - Added some colors to the state tree
+- Add collection type name to state tree
+    - Type can be clicked to open/close collection items
 - Fixes
 
 # 0.1.0 (2015-11-29)